requirements.txt
numpy==1.26.2
pandas==1.4.1
torch==1.11.0
nba_api==1.4.1
ipykernel==6.27.0
ipdb==0.13.9
tqdm==4.66.1
streamlit==1.28.2
PyYAML==6.0.1